Preparing to add simulation-to-general analysis transition.
If the top-level can't be completely simulated (if it does anything involving IO,
or pretty much anything a real program needs to) then it needs to switch to general
constant/type/etc analysis. However, that requires the top level be in SSA. Which
means SSA needs to be much, much faster. This commit includes many necessary speed
improvements as well as the simulation-level change. 2 tests don't work, as expected.