All (expected) specs pass!
All specs that passed before the switch to C++ basic blocks now pass. Bug fixed which made the difference: the "start block" of a LaserProc object could get changed by SSA, without the LaserProc realizing it. Since any number of transformations could also cause this bug, the "start block" is now calculated as the block-taken successor of the method call which bears the block.