Simulation gracefully terminates upon reaching unpredictable code.
This is the commit that moves Laser toward actually handling real-world
Ruby applications: it simulates as far as it can through predictable code,
and then switches to constant propagation. This requires SSA transformation
of *all* code, which was a performance killer, hence the C++ BasicBlock
implementation and massive performance boosts.