Added declare statements ala CL -> alias info @ rb_check_convert_type.
rb_check_convert_type is compiled to a couple of branches with a
result temp which holds the conversion. If that conversion is nil,
there's no clear alias path in the def/uses from SSA from the target
of the conversion to the result. Adding a :declare statement into
the IR allows this information to exist.

This corrects an issue with yield analysis; improved raise analysis
made clear the nil path which killed aliasing, thus killing inference.