Fixed conflicting tests for yield analysis.
Since the new 'warn about overriding block_given?' tests introduce a class which
overrides block_given?, the yield tests were noticing the existence of a block_given?
which was not Kernel#block_given?. Since the old yield tests were methods on Object,
and not on a particular class, they had to decide each method was block-optional.

By isolating each test in a class not equal to the ones introduced in the
override-warning specs, the yield tests no longer fail.